requirements.txt
fastapi==0.116.1
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
youtube_transcript_api==1.2.1
yt_dlp==2025.7.21
starlette==0.35.1
uvicorn==0.27.0
